infinite_moon_reader {xaringan}	R Documentation

Serve and live reload slides

Description

Use the servr package to serve and reload slides on change. inf_mr() is an alias of infinite_moon_reader().

Usage

infinite_moon_reader(moon, cast_from = ".")

inf_mr(moon, cast_from = ".")

Arguments

moon
The input Rmd file path (if missing and in RStudio, the current active document is used).

cast_from
The root directory of the server.

Details

The Rmd document is compiled continuously to trap the world in the Infinite Tsukuyomi. The genjutsu is cast from the directory specified by cast_from, and the Rinne Sharingan will be reflected off of the moon.

Note

This function is not really tied to the output format moon_reader(). You can use it to serve any single-HTML-file R Markdown output.

References

http://naruto.wikia.com/wiki/Infinite_Tsukuyomi

See Also

servr::httw

[Package xaringan version 0.9 Index]