summon_remark {xaringan}	R Documentation

Summon remark.js to your local disk

Description

Download a version of the remark.js script to your local disk, so you can render slides offline. You need to change the chakra argument of moon_reader() after downloading remark.js.

Usage

summon_remark(version = "latest", to = "libs/")

Arguments

version
The version of remark.js (e.g. latest, 0.13, or 0.14.1).

to
The destination directory.

[Package xaringan version 0.9 Index]